write universal srs to file moved to a helper file
